can: don't start listener before FD is initialized
Otherwise the can class might read from standard input (as the default
value of `s` is zero) which interferes with the uart. This is due to a
race condition since s is initialized after the thread has been started.

While at it also return from the constructor instead of setting stop.
Even better would probably be throwing an exception.