uart: Enforce size limit on rx_fifo and tx_fifo
Previously both fifos had an unlimited size even though the
specification (G000 Manual, chapter 19) limits the size to 8 bytes.

If this limit is not enforced the UART cannot reliably be used with
Linux as it only retrieves up to 8 byte per interrupt from it. This is
especially problematic with SLIP as the packets can't be processed in
time during high throughput.

Limits are enforced using a semaphores in background threads. Writes to
a full fifo are discarded (mandated by the specification).