iss: call quantum_keeper.sync() regardless of bus lock state
This resolves a race condition observed when using linux-vp with
multiple cores. This race condition would occur when a hart tried to
access a resource protected by a spinlock already held by a different
hart. In this case, the thread attempting to acquire the spinlock, would
cause a deadlock since it wouldn't return control to the SystemC
scheduler because the bus was locked by an LR_W instruction (part of the
spinlock code).