iss: check for breakpoints before doing run_step in exec_step
Sometimes GDB seems to set a breakpoint to the current PC. I am not
entirely sure why but I observed the following protocol exchange during
a next command:

	Sending packet: $z0,80000008,4#c6...Ack
	Packet received: OK
	Sending packet: $T1#85...Ack
	Packet received: OK
	Sending packet: $Hg1#e0...Ack
	Packet received: OK
	Sending packet: $p20#d2...Ack
	Packet received: 0400008000000000
	Sending packet: $m80000000,40#85...Ack
	Packet received: …
	Sending packet: $m80000004,4#59...Ack
	Packet received: b3840500
	Sending packet: $Z0,80000004,4#a2...Ack
	Packet received: OK

Here GDB removes the old breakpoint at 0x80000008, checks if thread 1 is
still alive and switches to it. Afterwards it retrieves the current PC
from thread 1 (80000004), reads the instruction from that address and
sets a breakpoint for it.

Without this change this breakpoint would never be hit. This is in
accordance with the GDB documentation [0] which states:

	 The breakpoint will stop your program just before it executes
	 any of the code in the specified location.

[0]: https://sourceware.org/gdb/onlinedocs/gdb/Set-Breaks.html