uart: Add input state machine to exit VP without sending signal
Since the VPs use the input TTY in raw mode, one cannot stop the VP by
pressing Ctrl-C as it will not emit a SIGINT. Similar to QEMU, this code
implements an input state machine with two input modes. The control
input mode can be entered by pressing Ctrl-a, in this mode an exit
command can be issued by pressing Ctrl-c. Pressing Ctrl-a twice will
send the Ctrl-a control character to the guest.